+++ analysis/bin/BuildFile.xml
@@ -6,14 +6,14 @@
   <lib name="EG"/>
   <use name="nano/analysis"/>
 </bin>
-<!--<bin name="topAnalysis" file="topAnalysis.cc">
+<bin name="massAnalysis" file="massAnalysis.cc">
   <use name="root"/>
   <use name="rootxml"/>
   <lib name="Eve"/>
   <lib name="EG"/>
   <use name="nano/analysis"/>
   <flags CXXFLAGS="-lTMVA"/>
-</bin>-->
+</bin>
 <bin name="vtsAnalysis" file="vtsAnalysis.cc">
   <use name="root"/>
   <use name="rootxml"/>
